Configure docker and docker-compose to bootstrap Rails project
Create Dockerfile, docker-compose.yml, Gemfile and Gemfile.lock
according to instruction found in [1] with only a few changes, such as
adding an 'ENTRYPOINT' and 'CMD' in the Dockerfile to set the default
action of containers created from this API image, and defining a volume
for the dtabase container in docker-compose.yml.

Create .env file to save environment variables for local development.

Create .dockerignore file to ignore files that shou;d not be copied to
the API image.

[1]: https://docs.docker.com/compose/rails/

Signed-off-by: Victor Hugo M. Pinto <victor2142@gmail.com>